Crate rpcap

This crate has functionality for reading and writing packet capture packet files in the traditional libpcap file format. There is no support for the newer pcapng file format.

Please note that there is no functionality for capturing live packets from a network interface, this library only handles reading/writing data in the pcap file format.

Modules

read
Functionality for reading a packet capture.

write
Functionality for writing packet captures.

Structs

CapturedPacket
The CapturedPacket struct contains information about a single captured packet.

Enums

Linktype
Known identifiers for the types of packets that might be captured in a pcap file. This tells you how to interpret the packets you receive.

PcapError
The error type for this crate.