Enum rpcap::Linktype

#[repr(u32)]

pub enum Linktype {
    NULL,
    ETHERNET,
    AX25,
    IEEE802_5,
    ARCNET_BSD,
    SLIP,
    PPP,
    FDDI,
    PPP_HDLC,
    PPP_ETHER,
    ATM_RFC1483,
    RAW,
    C_HDLC,
    IEEE802_11,
    FRELAY,
    LOOP,
    LINUX_SLL,
    LTALK,
    PFLOG,
    IEEE802_11_PRISM,
    IP_OVER_FC,
    SUNATM,
    IEEE802_11_RADIOTAP,
    ARCNET_LINUX,
    APPLE_IP_OVER_IEEE1394,
    MTP2_WITH_PHDR,
    MTP2,
    MTP3,
    SCCP,
    DOCSIS,
    LINUX_IRDA,
    USER00_LINKTYPE,
    USER01_LINKTYPE,
    USER02_LINKTYPE,
    USER03_LINKTYPE,
    USER04_LINKTYPE,
    USER05_LINKTYPE,
    USER06_LINKTYPE,
    USER07_LINKTYPE,
    USER08_LINKTYPE,
    USER09_LINKTYPE,
    USER10_LINKTYPE,
    USER11_LINKTYPE,
    USER12_LINKTYPE,
    USER13_LINKTYPE,
    USER14_LINKTYPE,
    USER15_LINKTYPE,
    IEEE802_11_AVS,
    BACNET_MS_TP,
    PPP_PPPD,
    GPRS_LLC,
    GPF_T,
    GPF_F,
    LINUX_LAPD,
    BLUETOOTH_HCI_H4,
    USB_LINUX,
    PPI,
    IEEE802_15_4,
    SITA,
    ERF,
    BLUETOOTH_HCI_H4_WITH_PHDR,
    AX25_KISS,
    LAPD,
    PPP_WITH_DIR,
    C_HDLC_WITH_DIR,
    FRELAY_WITH_DIR,
    IPMB_LINUX,
    IEEE802_15_4_NONASK_PHY,
    USB_LINUX_MMAPPED,
    FC_2,
    FC_2_WITH_FRAME_DELIMS,
    IPNET,
    CAN_SOCKETCAN,
    IPV4,
    IPV6,
    IEEE802_15_4_NOFCS,
    DBUS,
    DVB_CI,
    MUX27010,
    STANAG_5066_D_PDU,
    NFLOG,
    NETANALYZER,
    NETANALYZER_TRANSPARENT,
    IPOIB,
    MPEG_2_TS,
    NG40,
    NFC_LLCP,
    INFINIBAND,
    SCTP,
    USBPCAP,
    RTAC_SERIAL,
    BLUETOOTH_LE_LL,
    NETLINK,
    BLUETOOTH_LINUX_MONITOR,
    BLUETOOTH_BREDR_BB,
    BLUETOOTH_LE_LL_WITH_PHDR,
    PROFIBUS_DL,
    PKTAP,
    EPON,
    IPMI_HPM_2,
    ZWAVE_R1_R2,
    ZWAVE_R3,
    WATTSTOPPER_DLM,
    ISO_14443,
    RDS,
    USB_DARWIN,
}

Known identifiers for the types of packets that might be captured in a pcap file. This tells you how to interpret the packets you receive.

Look at tcpdump.org for the canonical list with descriptions.

Variants
NULLETHERNET

Ethernet packets

AX25IEEE802_5ARCNET_BSDSLIPPPPFDDIPPP_HDLCPPP_ETHERATM_RFC1483RAW

IP packets (IPv4 or IPv6)

C_HDLCIEEE802_11FRELAYLOOPLINUX_SLLLTALKPFLOGIEEE802_11_PRISMIP_OVER_FCSUNATMIEEE802_11_RADIOTAPARCNET_LINUXAPPLE_IP_OVER_IEEE1394MTP2_WITH_PHDRMTP2MTP3SCCPDOCSISLINUX_IRDAUSER00_LINKTYPEUSER01_LINKTYPEUSER02_LINKTYPEUSER03_LINKTYPEUSER04_LINKTYPEUSER05_LINKTYPEUSER06_LINKTYPEUSER07_LINKTYPEUSER08_LINKTYPEUSER09_LINKTYPEUSER10_LINKTYPEUSER11_LINKTYPEUSER12_LINKTYPEUSER13_LINKTYPEUSER14_LINKTYPEUSER15_LINKTYPEIEEE802_11_AVSBACNET_MS_TPPPP_PPPDGPRS_LLCGPF_TGPF_FLINUX_LAPDBLUETOOTH_HCI_H4USB_LINUXPPIIEEE802_15_4SITAERFBLUETOOTH_HCI_H4_WITH_PHDRAX25_KISSLAPDPPP_WITH_DIRC_HDLC_WITH_DIRFRELAY_WITH_DIRIPMB_LINUXIEEE802_15_4_NONASK_PHYUSB_LINUX_MMAPPEDFC_2FC_2_WITH_FRAME_DELIMSIPNETCAN_SOCKETCANIPV4IPV6IEEE802_15_4_NOFCSDBUSDVB_CIMUX27010STANAG_5066_D_PDUNFLOGNETANALYZERNETANALYZER_TRANSPARENTIPOIBMPEG_2_TSNG40NFC_LLCPINFINIBANDSCTPUSBPCAPRTAC_SERIALBLUETOOTH_LE_LLNETLINKBLUETOOTH_LINUX_MONITORBLUETOOTH_BREDR_BBBLUETOOTH_LE_LL_WITH_PHDRPROFIBUS_DLPKTAPEPONIPMI_HPM_2ZWAVE_R1_R2ZWAVE_R3WATTSTOPPER_DLMISO_14443RDSUSB_DARWIN