Struct rpcap::CapturedPacket

pub struct CapturedPacket<'a> {
    pub time: Timespec,
    pub data: &'a [u8],
    pub orig_len: usize,
}

The CapturedPacket struct contains information about a single captured packet.

Fields
time: Timespec

The time when the packet was captured.

data: &'a [u8]

The contents of the packet (possibly truncated to orig_len bytes during capture). Depending on the Linktype of the capture, there might be completely different data in this packet. The user of this library is responsible for interpreting the contents correctly.

orig_len: usize

The size of the packet as it was on the wire. Might be larger than the size of data, in which case data was truncated and is incomplete.